pyproject.toml
[tool.poetry]
name = "chibi"
version = "1.2.0"
description = "An asynchronous Telegram bot providing access to various LLMs (OpenAI, Gemini, Anthropic, etc.) and image generation models, featuring context management and built-in tools for web search, news retrieval, and reading web pages."
authors = ["Sergei Nagaev <nagaev.sv@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.38.3"
fake-useragent = "2.2.0"
httpx = "0.28.1"
loguru = "0.7.3"
openai = "1.77.0"
pydantic = "2.11.4"
pydantic-settings = "2.9.1"
python-telegram-bot = {extras = ["socks", "job-queue"], version = "^22.0"}
redis = {extras = ["hiredis"], version = "5.2.1"}
telegramify-markdown = "0.5.1"
trafilatura = "2.0.0"
aiocache = "0.12.3"
anthropic = "^0.72.1"
google-genai = "1.51.0"
ddgs = "^9.9.0"
dashscope = "^1.25.1"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.3"
fakeredis = "^2.28.1"
freezegun = "^1.5.1"
moto = "^5.1.4"
pytest = "^8.3.5"
respx = "^0.22.0"
ruff = "^0.11.6"
vulture = "^2.14"
mypy = "^1.15.0"
ipython = "^9.7.0"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
files = "**/*.py"
follow_imports = "silent"
ignore_missing_imports = true
show_column_numbers = true
strict = false
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
allow_untyped_calls = true
plugins = [ "pydantic.mypy" ]


[tool.vulture]
ignore_names = ["expected_type",]
min_confidence = 80
sort_by_size = true
paths = ["."]

[tool.ruff]
line-length = 120
exclude = [
    ".git",
    "__pycache__",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
